일단지형생성 되긴함 다시해봄!

+++ D3D/Client/Terrain.h
@@ -21,7 +21,7 @@ private:
 private:
 	BYTE* _HeightMapPixels;
 	vec3 _scale;
-	float _width;
-	float _length;
+	int _width;
+	int _length;
 };
